% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{Generate the indices to subset a dataset for cross validation}
\usage{
cross_validate(n, fold, folds = 10, random_order = TRUE, seed = 1)
}
\arguments{
\item{n}{Dataset size.}

\item{fold}{The i th fold.}

\item{folds}{Number of folds.}

\item{random_order}{Permute the order of the indices.}

\item{seed}{Sets the seed when permuting the order.}
}
\description{
Generate the indices to subset a dataset for cross validation
}
